:truck: Move heroes mocks from feature component to app root
